requirements.txt
Django==5.2.1
djangorestframework==3.16.0
drf-spectacular==0.28.0
requests==2.32.3
python-dotenv==1.1.0
pytrends==4.9.2
scikit-learn==1.6.1
numpy==2.2.5
scipy==1.15.3
psycopg2-binary==2.9.10
gunicorn==23.0.0
# Added based on previous requirements.txt and potential use
asgiref==3.8.1 # Django dependency
django-cors-headers==4.7.0 # Common for web apps
drf-yasg==1.21.10 # Alternative to drf-spectacular, often used together or one chosen
# Dependencies for pandas, numpy, scikit-learn if not covered by their direct install
python-dateutil # pandas dependency
pytz # pandas dependency
# Common utilities
six # often a dependency
sqlparse # Django dependency
tzdata # timezone support
uritemplate # drf-spectacular dependency
packaging # drf-spectacular dependency
PyYAML # drf-spectacular dependency
inflection # drf-spectacular dependency
# Keep whitenoise as it's common for serving static files with Django
whitenoise==6.9.0
